+++ NEWS.md
@@ -9,6 +9,7 @@
 ## New features
 
 - Add vignette "Reading and Displaying inForm Image Files"
+- Add vignette "Computing Inter-cellular Distances"
 - `spatial_distribution_report` creates an HTML report showing the 
   location and nearest-neighbor relations of cells in a single field.
 - `read_components` reads component image files.